♻️[#157] 뷰 분리
